+++ test/pallet/crate/lein_test.clj
@@ -1,23 +1,26 @@
 (ns pallet.crate.lein-test
   (:require
    [clojure.test :refer :all]
+   [pallet.action :refer [with-action-options]]
    [pallet.actions :refer [exec-checked-script remote-file]]
+   [pallet.api :refer [plan-fn server-spec]]
    [pallet.build-actions :refer [build-actions]]
-   [pallet.crate.lein :refer :all]
+   [pallet.crate.lein :as lein]
+   [pallet.crate.java :as java]
    [pallet.test-utils]))
 
 (deftest install-lein-test
   (is (script-no-comment=
-       (first (build-actions {:phase-context "install-lein"}
+       (first (build-actions {:phase-context "install"}
                 (remote-file
                  "/usr/local/bin/lein"
-                 :url (format *lein-url* "stable")
+                 :url (format lein/*lein-url* "stable")
                  :insecure true
                  :no-versioning true
                  :mode "755")))
        (first  (build-actions {}
-                 (lein-settings {})
-                 (install-lein))))))
+                 (lein/settings {})
+                 (lein/install {}))))))
 
 (deftest lein-test
   (is (script-no-comment=
@@ -26,8 +29,15 @@
                  "lein test"
                  ("/usr/local/bin/lein" test))))
        (first  (build-actions {}
-                 (lein-settings {})
-                 (lein :test))))))
+                 (lein/settings {})
+                 (lein/lein :test))))))
 
 (deftest leiningen-test
-  (is (leiningen {})))
+  (is (lein/server-spec {})))
+
+(def live-test-spec
+  (server-spec
+   :extends [(lein/server-spec {}) (java/server-spec {})]
+   :phases {:test (plan-fn
+                    (with-action-options {:script-prefix :no-sudo}
+                      (lein/lein "version")))}))