+++ ReleaseNotes.md
@@ -1,3 +1,9 @@
+## 0.8.0-alpha.2
+
+- Make clojure and pallet dependencies :provided scope
+
+- Use standard names for plan functions
+
 # Release Notes
 
 ## 0.8.0-alpha.1